requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
greenlet==3.1.1             # required by SQLAlchemy async on Python 3.13
psycopg2-binary==2.9.9
pgvector==0.3.2

# Job queue
redis==5.0.8
rq==1.16.2

# GitHub API client
httpx==0.27.2
httpx-sse==0.4.0            # MCP SSE transport

# AST parsing
tree-sitter==0.23.1
tree-sitter-python==0.23.2
tree-sitter-javascript==0.23.0
tree-sitter-typescript==0.23.0
tree-sitter-java==0.23.2
tree-sitter-go==0.23.1
tree-sitter-rust==0.23.0

# Embeddings & ML
voyageai==0.3.0
sentence-transformers         # cross-encoder reranker (version pinned by torch)

# Token counting
tiktoken==0.7.0

# Config & validation
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# MCP
mcp[cli]>=1.26.0

# Auth
PyJWT>=2.9.0
cryptography>=43.0.1

# Admin UI
streamlit>=1.38.0
plotly>=5.0.0

# Utilities
python-multipart==0.0.9       # FastAPI form data
structlog==24.4.0             # structured logging
